job_message/requirements.txt
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.2
fake-useragent==0.1.11
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
lxml==4.5.0
mysql==0.0.2
mysqlclient==1.4.6
parsel==1.5.2
parso==0.6.2
pickleshare==0.7.5
prompt-toolkit==3.0.4
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pywin32==227
queuelib==1.5.0
requests==2.23.0
Scrapy==2.0.0
scrapy-fake-useragent==1.2.0
scrapyd==1.2.1
scrapyd-client==1.1.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
traitlets==4.3.3
Twisted==19.10.0
urllib3==1.25.8
w3lib==1.21.0
wcwidth==0.1.8
zope.interface==4.7.2
